fix(config): unify prompt and trace output projections
